+++ app/views/content/index.html.erb
@@ -1,6 +1,12 @@
-<router-view v-slot="{ Component, route }">
-  <!-- Use any custom transition and fallback to `fade` -->
-  <transition :name="route.meta.transition || 'fade'">
-    <component :is="Component" />
-  </transition>
-</router-view>
+<div data-behavior="vue" class="flex flex-col h-screen">
+  <reload-pwa></reload-pwa>
+  <app-header></app-header>
+  <main class="bg-white flex-grow overflow-auto dark:bg-gray-800 dark:text-gray-200">
+    <router-view v-slot="{ Component, route }">
+      <!-- Use any custom transition and fallback to `fade` -->
+      <transition :name="route.meta.transition || 'fade'">
+        <component :is="Component" />
+      </transition>
+    </router-view>
+  </main>
+</div>
